% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_stationv.R
\name{flow_stationv}
\alias{flow_stationv}
\title{Predict Mean Daily Flow}
\usage{
flow_stationv(station, date_vector, reference = NA)
}
\arguments{
\item{station}{A single MLocID for an ODEQ monitoring station (example: “UmpNF-078”, "31878-ORDEQ", or "14138900")}

\item{date_vector}{A vector of dates in “YYYY-MM-DD” format, not necessarily continuous}

\item{reference}{(optional) A USGS stream gage ID to use as the reference. If not provided, the nearest will be automatically selected.}
}
\value{
A list object with: 1 = a dataframe with the mean daily flow series and 2 = additional details
}
\description{
Estimates mean daily flow series for ungaged stream site in Oregon,using USGS StreamStats and NWIS data. Based on the methods described in Lorenz & Ziegeweid (2016) and other USGS publications.
}
\details{
Returns a list with two dataframes. The first dataframe (1) is the generated mean daily flow timeseries. The second dataframe (2) contains additional details about the request, such as:
\itemize{
\item The target site longitude and latitude
\item The USGS reference gage used
\item The drainage area of the watershed delineated by StreamStats for the target site
}

Note: If the returned drainage area is zero or much smaller than expected, check the StreamStats website to ensure that the target coordinates align with the stream grid. (https://streamstats.usgs.gov/ss/)
}
\examples{
# If no reference gage is entered, the nearest will be automatically chosen:
result <- flow_stationv("UmpNF-078",c("2010-01-01","2011-03-04","2012-12-31"))
timeseries <- result[[1]]
flows <- timeseries$`PredictedFlow(cfs)`
details <- result[[2]]

# The reference gage can be specified, if desired:
result <- flow_stationv("UmpNF-078",c("2010-01-01","2011-03-04","2012-12-31"), 14318000)

}
