% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.Mean.Daily.Flow.Coord_Series.R
\name{Predict.Mean.Daily.Flow.Coord_Series}
\alias{Predict.Mean.Daily.Flow.Coord_Series}
\title{Predict Mean Daily Flow}
\usage{
Predict.Mean.Daily.Flow.Coord_Series(latitude, longitude, start_date, end_date)
}
\arguments{
\item{latitude}{The latitude of a target stream site, in decimal degrees}

\item{longitude}{The longitude of a target stream site, in decimal degrees}

\item{start_date}{A single date in “YYYY-MM-DD” format for the start of a continuous daily series}

\item{end_date}{A single date in “YYYY-MM-DD” format for the end of a continuous daily series}
}
\value{
A list object with: 1 = a dataframe with the mean daily flow series and 2 = additional details
}
\description{
Estimates mean daily flow series for ungaged stream site in Oregon,using USGS StreamStats and NWIS data. Based on the methods described in Lorenz & Ziegeweid (2016) and other USGS publications.
}
\details{
Returns a list with two dataframes. The first dataframe (1) is the generated mean daily flow timeseries. The second dataframe (2) contains additional details about the request, such as:
\itemize{
\item The target site longitude and latitude
\item The nearest USGS reference gage identified
\item The drainage area of the watershed delineated by StreamStats for the target site
}

Note: If the returned drainage area is zero or much smaller than expected, check the StreamStats website to ensure that the target coordinates align with the stream grid. (https://streamstats.usgs.gov/ss/)
}
\examples{
result <- Predict.Mean.Daily.Flow.Coord_Series(43.252618,-123.026172,"2010-01-01","2012-12-31")
timeseries <- result[[1]]
flows <- timeseries$`PredictedFlow(cfs)`
details <- result[[2]]

}
